Fixing dash-as-sh problems with install.sh
The read builtin in dash behaves differently than any other read builtin
I've encountered, so we have to pass it a junk variable despite us never
planning to use it.  It seems like $REPLY isn't getting the results of a
read command if you don't give it a parameter.  Weird.  I could go figure
out what the correct behaviour is, but I'm honestly not that interested.

Signed-off-by: Joe MacDonald <joe@deserted.net>