Adding gmail / imap support
I do prefer to send my patches via gmail since that's typically where I
any of my project-related email goes anyway and on at least some machines
it saves me from having to set up an MTA.

This also adds in a couple of reasonably essential options, too.  The
default behaviour to not suppress cc-ing everyone in the signoff chain is
just plain wrong, for example, and giggle is basically a less-ugly gitk.

Signed-off-by: Joe MacDonald <joe@deserted.net>