Command-line based dictionary lookups
I'd been using curl to look up stuff from dict.org for a while but I
dislike that it only knows how to either find me a definition or do a
similarity match on a word.  So these additions let you look up a
definition if you get an exact match, otherwise you get a list of possible
matches for your random jumble of letters.

Signed-off-by: Joe MacDonald <joe@deserted.net>