Fixing function definitions problems on ksh93
Tripped over a new version of ksh93 yesterday:

   Version M 93s+ 2008-01-31

which seems to be the first one that barfed on my fns.wrs function
definitions.  Weird, but whatever.  The only one that was causing problems
was a function that I only ever intended to be internal anyway, so no
great loss in readability.

Signed-off-by: Joe MacDonald <joe@deserted.net>