Add a configure function to pass all the right parameters
Remembering paths is a giant pain.  I used to have a wrcfg function way
back in the day but it's long ago disappeared because everything changed.
This one's hopefully a little more likely to survive, but who knows.
Works for now and I don't have to remember where the prebuilt RPMs live.

Signed-off-by: Joe MacDonald <joe@deserted.net>