Updating git aliases
'git br' will now show branches but only the one you're currently on and a
handful of the ones around you.  That may or may not be useful, but I've
found it to be helpful more often than not.

Most of the other !sh-based aliases were broken, had been for some time
and I just hadn't noticed, so I fixed those too.  And swept away the ones
that weren't broken because shelling out is just plain wrong anyway.
Shell functions work, learn to use them!

Signed-off-by: Joe MacDonald <joe.macdonald@windriver.com>