sh: add purgelist file processing function
I'm tired of trying to figure out how to delete trees of files with messed
up names, so I wrote this function that will read a file that is a list of
files to delete and unlink them without any of the shell escape nonsense.
Seems to not fall over when I was cleaning out a pile of mp3 files with
weird characters in the names, so I'm sure it's gold.

Signed-off-by: Joe MacDonald <joe.macdonald@windriver.com>