misc: shell script to clean up .u1conflict files
I love Ubuntu One.  I use it a lot.  From my work laptop, my netbook and
home home desktop.  After a while I end up with a *lot* of .u1conflict
files that need cleaning up, so I wrote this script to ease that process
for me.  it take exactly one (optional) option, -d, which is the directory
to start searching in.

Signed-off-by: Joe MacDonald <joe.macdonald@windriver.com>