sh: Add support for 256-color terminals
Fixing up support for setting TERM to xterm-256color and screen-256color
where appropriate, working around a bug in XFCE's terminal application at
the same time.  This is really only useful if you're using console vim,
mutt with the expanded colour scheme added here, or with a set of
dircolors with make use of the full colour range.  For example, one
generated from:

   git://github.com/joeythesaint/dotfiles-dircolors.git

Signed-off-by: Joe MacDonald <joe.macdonald@windriver.com>