dustbin: add ripper script
The intent was to have a script I could use to pull episodes off a DVD and
have them named and numbered in a way I like with minimal interaction.
Similar to the behaviour from various CD ripping programs out there (like
Sound Juicer).  Never quite worked because of the crazy variety possible
in DVD structures.

Signed-off-by: Joe MacDonald <joe@deserted.net>